Analytics/Data handler/requirements.txt
flask==1.1.1
influxdb-client==1.43.0
numpy==1.19.5            # works with Python 3.9 and pandas 1.2.3
pandas==1.2.3
tqdm==4.42.1
pyyaml==5.4
requests==2.31.0         # <-- bump to resolve urllib3 conflict (2.32.x also OK)
jinja2<3.1.0
itsdangerous==2.0.1
werkzeug==2.0.3
